aggiunta back to top button, fixati bug in aggiunta e dao per ritorno double rating, aggiunte classi failure e successful per messaggi, aggiunta logica e presentazione form di modifica, modifica script js per il caricamento immagine in form aggiunta e modifica

+++ src/main/java/model/ProcessorDAO.java
@@ -17,7 +17,7 @@ public class ProcessorDAO {
                 Processor processor = new Processor();
                 processor.setId(resultSet.getInt("id"));
                 processor.setName(resultSet.getString("name"));
-                processor.setRating(resultSet.getInt("rating"));
+                processor.setRating(resultSet.getDouble("rating"));
                 processor.setPrice(resultSet.getDouble("price"));
                 processor.setShop_URL(resultSet.getString("shop_URL"));
                 processor.setImage_URL(resultSet.getString("image_URL"));
